Move GetIPFPath into ExperimentInfo
Moves the method for getting IPF Paths into Experiment info as a static
method, so it can be called internally and by Python APIs
