+++ Framework/API/src/ExperimentInfo.cpp
@@ -43,7 +43,6 @@
 #include <boost/algorithm/string.hpp>
 #include <boost/lexical_cast.hpp>
 #include <boost/regex.hpp>
-#include <memory>
 
 #include <Poco/DirectoryIterator.h>
 #include <Poco/Path.h>
@@ -51,6 +50,9 @@
 #include <Poco/SAX/ContentHandler.h>
 #include <Poco/SAX/SAXParser.h>
 #include <nexus/NeXusException.hpp>
+
+#include <algorithm>
+#include <memory>
 #include <tuple>
 
 using namespace Mantid::Geometry;
@@ -1002,6 +1004,85 @@ ExperimentInfo::getInstrumentFilename(const std::string &instrumentName,
     g_log.debug() << "No instrument file found\n";
   }
   return instFile;
+}
+
+//-----------------------------------------------------------------------------------------------------------------------
+/// Search the directory for the Parameter IDF file and return full path name if
+/// found, else return "".
+//  directoryName must include a final '/'.
+std::string ExperimentInfo::getFullPathParamIDF(std::string instName,
+                                                const std::string &dirHint) {
+
+  constexpr auto lookupFile = [](const std::string &dir,
+                                 const std::string &filename) {
+    Poco::Path directoryPath(dir);
+    directoryPath.makeDirectory();
+    // Remove the path from the filename
+    Poco::Path filePath(filename);
+    const std::string &instrumentFile = filePath.getFileName();
+
+    // First check whether there is a parameter file whose name is the same as
+    // the IDF file, but with 'Parameters' instead of 'Definition'.
+    std::string definitionPart("_Definition");
+    const std::string::size_type prefix_end(
+        instrumentFile.find(definitionPart));
+    const std::string::size_type suffix_start =
+        prefix_end + definitionPart.length();
+    // Get prefix and leave case sensitive
+    std::string prefix = instrumentFile.substr(0, prefix_end);
+    // Make suffix ensuring it has positive length
+    std::string suffix = ".xml";
+    if (suffix_start < instrumentFile.length()) {
+      suffix = instrumentFile.substr(suffix_start, std::string::npos);
+    }
+
+    // Assemble parameter file name
+    std::string fullPathParamIDF =
+        directoryPath.setFileName(prefix + "_Parameters" + suffix).toString();
+    if (!Poco::File(fullPathParamIDF)
+             .exists()) { // No such file exists, so look
+                          // for file based on instrument
+                          // ID
+                          // given by the prefix
+      fullPathParamIDF =
+          directoryPath.setFileName(prefix + "_Parameters.xml").toString();
+    }
+
+    if (!Poco::File(fullPathParamIDF)
+             .exists()) { // No such file exists, indicate
+                          // none found in this directory.
+      fullPathParamIDF = "";
+    }
+
+    return fullPathParamIDF;
+  };
+
+  // All inst names are stored as capitals currently,
+  // so we need to also do this for case-sensitive filesystems
+  std::transform(instName.begin(), instName.end(), instName.begin(), ::toupper);
+
+  // Try the hinted dir first
+  if (!dirHint.empty()) {
+    const std::string result = lookupFile(dirHint, instName);
+    if (!result.empty()) {
+      return result;
+    }
+  }
+
+  Kernel::ConfigServiceImpl &configService = Kernel::ConfigService::Instance();
+  std::vector<std::string> directoryNames =
+      configService.getInstrumentDirectories();
+
+  for (const auto &dirName : directoryNames) {
+    // This will iterate around the directories from user ->etc ->install, and
+    // find the first beat file
+    const std::string result = lookupFile(dirName, instName);
+    if (!result.empty()) {
+      return result;
+    }
+  }
+
+  return "";
 }
 
 /** Return a const reference to the DetectorInfo object.
